fix d tag crash

+++ storage/postgresql/save.go
@@ -19,8 +19,10 @@ func (b *PostgresBackend) SaveEvent(evt *nostr.Event) error {
 	} else if evt.Kind >= 30000 && evt.Kind < 40000 {
 		// NIP-33
 		d := evt.Tags.GetFirst([]string{"d"})
-		b.DB.Exec(`DELETE FROM event WHERE pubkey = $1 AND kind = $2 AND created_at < $3 AND tagvalues && ARRAY[$4]`,
-			evt.PubKey, evt.Kind, evt.CreatedAt.Unix(), d.Value())
+		if d != nil {
+			b.DB.Exec(`DELETE FROM event WHERE pubkey = $1 AND kind = $2 AND created_at < $3 AND tagvalues && ARRAY[$4]`,
+				evt.PubKey, evt.Kind, evt.CreatedAt.Unix(), d.Value())
+		}
 	}
 
 	// insert
